code-review cleanup: move plane-parsing to shared module, dedup, cache env read
- Move DataPlaneCloudProvider and parse_data_plane_name out of the
  server::public::graphql module into crate::data_plane, alongside the
  other plane-domain helpers. Directive/provisioning logic (beta_onboard)
  reaching into a GraphQL-presentation module for domain parsing was a
  layering smell; graphql::data_planes.rs and create_data_plane.rs now
  both consume the shared home instead.
- Add PUBLIC_DATA_PLANE_PREFIX and is_selectable_public_plane() so the
  "public, non-deprecated plane" predicate has one Rust-side definition
  instead of being re-expressed inline in the agent; the prefix is also
  bound as a SQL parameter in provision_tenant's public_planes query
  instead of a duplicated string literal.
- Cache COLOCATED_TRIAL_BUCKETS behind a OnceLock so the gate is read
  from the environment once per process instead of on every signup.
- Reorder Cargo.toml's sha2/hex to match the file's existing alphabetical
  dependency ordering.

No behavior change: all existing unit/integration tests pass unmodified,
verified against a real migrated Postgres.
